added service for ABM of buses

+++ src/main/java/ar/edu/unq/tip/model/Position.java
@@ -3,29 +3,29 @@ package ar.edu.unq.tip.model;
 public class Position extends Entity {
 
 	private static final long serialVersionUID = 296870812500050872L;
-	private float lat;
-	private float lng;
+	private double lat;
+	private double lng;
 
 	public Position() { }
 
-	public Position(float lat, float lng) {
+	public Position(double lat, double lng) {
 		this.lat = lat;
 		this.lng = lng;
 	}
 	
-	public float getLat() {
+	public double getLat() {
 		return lat;
 	}
 	
-	public void setLat(float lat) {
+	public void setLat(double lat) {
 		this.lat = lat;
 	}
 	
-	public float getLng() {
+	public double getLng() {
 		return lng;
 	}
 	
-	public void setLng(float lng) {
+	public void setLng(double lng) {
 		this.lng = lng;
 	}	
 }
